bq2snowflake/requirements.txt
asn1crypto==0.24.0
azure-common==1.1.23
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
boto3==1.9.235
botocore==1.12.235
cachetools==3.1.1
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
docutils==0.15.2
future==0.17.1
google-api-core==1.14.2
google-auth==1.6.3
google-cloud-bigquery==1.20.0
google-cloud-core==1.0.3
google-resumable-media==0.4.1
googleapis-common-protos==1.6.0
idna==2.8
ijson==2.5
jmespath==0.9.4
protobuf==3.9.2
pyasn1==0.4.7
pyasn1-modules==0.2.6
pycparser==2.19
pycryptodomex==3.9.0
PyJWT==1.7.1
pyOpenSSL==19.0.0
python-dateutil==2.8.0
pytz==2019.2
requests==2.22.0
rsa==4.0
s3transfer==0.2.1
six==1.12.0
snowflake-connector-python==1.9.0
urllib3==1.25.6
